backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
paddleocr==2.7.3
opencv-python-headless==4.10.0.84
pymupdf==1.24.11
pillow==10.4.0
pytesseract==0.3.13
openpyxl==3.1.5
numpy==2.3.5
